+++ src/appmixer/intercom/artifacts/test/GetContact.test.js
@@ -1,12 +1,11 @@
 const path = require('path');
-require('dotenv').config({ path: path.join(__dirname, '../.env') });
+require('dotenv').config({ path: path.join(__dirname, '../../../../../test/.env') });
 const assert = require('assert');
 
-describe('GetCompany Component', function() {
+describe('GetContact Component', function() {
     let context;
-    let GetCompany;
-    let testCompanyId;
-    let testIntercomId;
+    let GetContact;
+    let testContactId;
 
     this.timeout(30000);
 
@@ -18,7 +17,7 @@ describe('GetCompany Component', function() {
         }
 
         // Load the component
-        GetCompany = require(path.join(__dirname, '../../src/appmixer/intercom/core/GetCompany/GetCompany.js'));
+        GetContact = require('../../core/GetContact/GetContact.js');
 
         // Mock context
         context = {
@@ -47,42 +46,41 @@ describe('GetCompany Component', function() {
             }
         };
 
-        // Create a test company first
+        // Create a test contact first
         try {
-            testCompanyId = `test-get-company-${Date.now()}`;
-            const CreateUpdateCompany = require(path.join(__dirname, '../../src/appmixer/intercom/core/CreateUpdateCompany/CreateUpdateCompany.js'));
+            const CreateContact = require('../../core/CreateContact/CreateContact.js');
 
             const createContext = { ...context };
             createContext.messages = {
                 in: {
                     content: {
-                        company_id: testCompanyId,
-                        name: 'Test Company for Get'
+                        email: `test-get-contact-${Date.now()}@example.com`,
+                        name: 'Test Contact for Get'
                     }
                 }
             };
 
-            const createResult = await CreateUpdateCompany.receive(createContext);
-            testIntercomId = createResult.data.id; // Store the Intercom ID
+            const result = await CreateContact.receive(createContext);
+            testContactId = result.data.id;
         } catch (error) {
-            console.error('Error creating test company:', error.response?.data || error.message);
+            console.error('Error creating test contact:', error.response?.data || error.message);
             throw error;
         }
     });
 
-    it('should retrieve a company by id', async function() {
+    it('should retrieve a contact by id', async function() {
         context.messages.in.content = {
-            id: testIntercomId
+            id: testContactId
         };
 
         try {
-            const result = await GetCompany.receive(context);
+            const result = await GetContact.receive(context);
 
             assert(result, 'Should return a result');
-            assert(result.data, 'Should return company data');
-            assert(result.data.company_id === testCompanyId, 'Should return correct company');
+            assert(result.data, 'Should return contact data');
+            assert(result.data.id === testContactId, 'Should return correct contact');
         } catch (error) {
-            console.error('Error retrieving company:', error.response?.data || error.message);
+            console.error('Error retrieving contact:', error.response?.data || error.message);
             throw error;
         }
     });
@@ -91,24 +89,24 @@ describe('GetCompany Component', function() {
         context.messages.in.content = {};
 
         try {
-            await GetCompany.receive(context);
+            await GetContact.receive(context);
             assert.fail('Should have thrown an error');
         } catch (error) {
             assert(error.name === 'CancelError', 'Should throw CancelError');
-            assert(error.message.includes('Company ID is required'), 'Should have appropriate error message');
+            assert(error.message.includes('Contact ID is required'), 'Should have appropriate error message');
         }
     });
 
-    it('should handle non-existent company id gracefully', async function() {
+    it('should handle non-existent contact id gracefully', async function() {
         context.messages.in.content = {
-            id: 'non-existent-company-12345'
+            id: 'non-existent-contact-12345'
         };
 
         try {
-            await GetCompany.receive(context);
-            assert.fail('Should have thrown an error for non-existent company');
+            await GetContact.receive(context);
+            assert.fail('Should have thrown an error for non-existent contact');
         } catch (error) {
-            assert(error.response && error.response.status === 404, 'Should return 404 for non-existent company');
+            assert(error.response && error.response.status === 404, 'Should return 404 for non-existent contact');
         }
     });
 });
